Move menu styling from reusable wrapper to component; move Menu.tsx from project folder to common dir
